Added -keep_bin to avoid the delete and overwrite of the precompiled folder

+++ libs/SOURCE_driver.pm
@@ -106,11 +106,13 @@ if ($compile) {
 # BIN's
 if ("$precompiled_is_run" eq "yes") {
  $conf_bin  = "bin-precompiled-$ROBOT_string";
- chdir $BRANCH;
- &command("rm -fr $conf_bin; cp -fr bin $conf_bin");
- if (-d "lib/bin" ) {&command("cp lib/bin/* $conf_bin/")};
- if (-d "bin-libs") {&command("cp bin-libs/* $conf_bin/")};
- chdir $suite_dir;
+ if (not $keep_bin) {
+  chdir $BRANCH;
+  &command("rm -fr $conf_bin; cp -fr bin $conf_bin");
+  if (-d "lib/bin" ) {&command("cp lib/bin/* $conf_bin/")};
+  if (-d "bin-libs") {&command("cp bin-libs/* $conf_bin/")};
+  chdir $suite_dir;
+ }
 }else{
  $conf_bin  = "bin-$conf_file-$FC_kind-$ROBOT_string";
 }
